light IDX can use Mmap instead of File
I thought an Mmap allocates memory when it reads, but i guess it doesn't? or somehow its smart enough to know when to allocate, and when not to? In any case, i noticed its a lot faster with an Mmap than using File directly, and it still uses very little allocations