uses tiny vec when reading raw git objects
this is a rebase conflict resolution. in addition to using tiny vec for reading raw git objects,
this commit changes the decoding API to use a mutable decompressor reference to avoid re-allocating
a decompressor each time. This currently doesn't actually change much because the decompressor is just
allocated higher up anyway, but this API will be used in the future when we actually will retain
a decompressor state. Original commit description below:

massive performance boost here. most of our raw object reads can fit on the stack, so we say if a raw object is less than 2kb we put it on the stack, but otherwise we heap allocate. TinyVec handles all of that for us behind the scenes