adds two oid map implementations, see note:
an oid map can map an oid to any type T. its useful for example when you traverse a lot of Oids and you want to remember what you have parsed before, and map it to some data, for example a location of where it exists. both methods are very space efficient. The OidMap_trunc takes a truncated OidFull (ie: a 16 byte array), whereas the OidMap_u128 takes a u128 value as the key