adds log-oneline example, see note:
I spent 2 weeks on this log-oneline example, and it is still not as fast as git log --oneline, while
taking several shortcuts to try to be faster than git... That said, it can log the linux kernel
history in under 1 second (on my machine), and uses about half of the memory that git uses. I consider that
a mild success although ideally I'd like to be able to be faster than git. However, for most smaller repositories
it is just as fast as git (or at least not noticeably slower because both git and this example are so fast that the
bash builtin `time` cannot accurately report their run time).

I spent a lot of time profiling trying to improve it, and I made many improvements, but overall I think the issue
comes down to:
- storing/inserting objects into the map
- parsing merge tags
- some startup issues

the first two can be improved by carefully optimizing the OidMap.insert, and the parse_mergetag functions respectively, but
the third is still a mystery to me. If you run the example with:
`./target/release/log-oneline .git/objects <some_sha_hash> 1` then it will only log 1 commit, but it does this extermely slower
than what git would do. I think I am doing something in the beginning initialization that is
slowing everything down substantially.

I didn't want to look at git's source code because I considered that cheating, but I think if i wanted to finish
optimizing this to it's full potential, then I'd need to look at git's source to see if I am doing something totally wrong...

All that being said: the point of this library isn't to provide high level traversal of git objects anyway.
This library's purpose is to provide the low level functions that let you easily extract/view git objects
without having to ask the git cli for them.